Transfer-Encoding takes precedence over Content-Length.
This patch changes the way how misbehaving clients, that send
Content-Length together with Transfer-Encoding, are handled.
Currently we don't do dechunking but apply the processing
for messages with valid Content-Length. Unfortunately, this
brings significant inconsistency with S3 as its affects AWSv4.
The fix brings the same mechanism we have in the master branch
of ceph/civetweb repo.

Fixes: http://tracker.ceph.com/issues/21015
Signed-off-by: Radoslaw Zarzynski <rzarzyns@redhat.com>